app/requirements.txt
Telethon
psycopg2-binary
Flask
requests
schedule